Use more responsive layouts
