+++ auth/composables/src/debug/java/com/google/android/horologist/auth/composables/screens/SelectAccountScreenPreview.kt
@@ -19,7 +19,6 @@ package com.google.android.horologist.auth.composables.screens
 import androidx.compose.runtime.Composable
 import androidx.wear.compose.ui.tooling.preview.WearPreviewDevices
 import com.google.android.horologist.auth.composables.model.AccountUiModel
-import com.google.android.horologist.compose.layout.rememberResponsiveColumnState
 
 @WearPreviewDevices
 @Composable
@@ -30,7 +29,6 @@ fun SelectAccountScreenPreview() {
             AccountUiModel(email = "thisisaverylongemail@example.com"),
         ),
         onAccountClicked = { _, _ -> },
-        columnState = rememberResponsiveColumnState(),
     )
 }
 
@@ -43,7 +41,6 @@ fun SelectAccountScreenPreviewNoAvatar() {
             AccountUiModel(email = "thisisaverylongemailaccountsample@example.com"),
         ),
         onAccountClicked = { _, _ -> },
-        columnState = rememberResponsiveColumnState(),
         defaultAvatar = null,
     )
 }
