+++ auth/composables/src/main/java/com/google/android/horologist/auth/composables/screens/SelectAccountScreen.kt
@@ -30,7 +30,9 @@ import com.google.android.horologist.auth.composables.R
 import com.google.android.horologist.auth.composables.chips.AccountChip
 import com.google.android.horologist.auth.composables.model.AccountUiModel
 import com.google.android.horologist.compose.layout.ScalingLazyColumn
-import com.google.android.horologist.compose.layout.ScalingLazyColumnState
+import com.google.android.horologist.compose.layout.ScalingLazyColumnDefaults.ItemType
+import com.google.android.horologist.compose.layout.ScalingLazyColumnDefaults.padding
+import com.google.android.horologist.compose.layout.ScreenScaffold
 import com.google.android.horologist.compose.layout.rememberResponsiveColumnState
 import com.google.android.horologist.compose.material.Title
 import com.google.android.horologist.images.base.paintable.ImageVectorPaintable.Companion.asPaintable
@@ -46,35 +48,43 @@ public fun SelectAccountScreen(
     accounts: List<AccountUiModel>,
     onAccountClicked: (index: Int, account: AccountUiModel) -> Unit,
     modifier: Modifier = Modifier,
-    columnState: ScalingLazyColumnState = rememberResponsiveColumnState(),
     title: String = stringResource(id = R.string.horologist_select_account_title),
     defaultAvatar: Paintable? = Icons.Default.AccountCircle.asPaintable(),
 ) {
-    ScalingLazyColumn(
-        modifier = modifier,
-        columnState = columnState,
-    ) {
-        item { Title(title, Modifier.padding(bottom = 8.dp)) }
+    val columnState = rememberResponsiveColumnState(
+        contentPadding = padding(
+            first = ItemType.Text,
+            last = ItemType.Chip,
+        ),
+    )
 
-        items(accounts.size) { index ->
-            val account = accounts[index]
-            MaterialTheme(
-                typography = MaterialTheme.typography.copy(
-                    button = MaterialTheme.typography.button.copy(
-                        lineBreak = LineBreak(
-                            strategy = LineBreak.Strategy.Balanced,
-                            strictness = LineBreak.Strictness.Normal,
-                            wordBreak = LineBreak.WordBreak.Default,
+    ScreenScaffold(scrollState = columnState) {
+        ScalingLazyColumn(
+            modifier = modifier,
+            columnState = columnState,
+        ) {
+            item { Title(title, Modifier.padding(bottom = 8.dp)) }
+
+            items(accounts.size) { index ->
+                val account = accounts[index]
+                MaterialTheme(
+                    typography = MaterialTheme.typography.copy(
+                        button = MaterialTheme.typography.button.copy(
+                            lineBreak = LineBreak(
+                                strategy = LineBreak.Strategy.Balanced,
+                                strictness = LineBreak.Strictness.Normal,
+                                wordBreak = LineBreak.WordBreak.Default,
+                            ),
                         ),
                     ),
-                ),
-            ) {
-                AccountChip(
-                    account = account,
-                    onClick = { onAccountClicked(index, account) },
-                    colors = ChipDefaults.secondaryChipColors(),
-                    defaultAvatar = defaultAvatar,
-                )
+                ) {
+                    AccountChip(
+                        account = account,
+                        onClick = { onAccountClicked(index, account) },
+                        colors = ChipDefaults.secondaryChipColors(),
+                        defaultAvatar = defaultAvatar,
+                    )
+                }
             }
         }
     }
