+++ composables/src/test/kotlin/com/google/android/horologist/composables/SectionedListHeaderFooterTest.kt
@@ -24,9 +24,11 @@ import com.google.android.horologist.composables.SectionedListTest.Companion.Dow
 import com.google.android.horologist.composables.SectionedListTest.Companion.DownloadsHeader
 import com.google.android.horologist.composables.SectionedListTest.Companion.DownloadsLoaded
 import com.google.android.horologist.composables.SectionedListTest.Companion.DownloadsLoading
-import com.google.android.horologist.composables.SectionedListTest.Companion.SectionedListPreview
 import com.google.android.horologist.composables.SectionedListTest.Companion.downloads
-import com.google.android.horologist.compose.layout.ScalingLazyColumnDefaults
+import com.google.android.horologist.compose.layout.ScalingLazyColumnDefaults.ItemType
+import com.google.android.horologist.compose.layout.ScalingLazyColumnDefaults.padding
+import com.google.android.horologist.compose.layout.ScreenScaffold
+import com.google.android.horologist.compose.layout.rememberResponsiveColumnState
 import com.google.android.horologist.screenshots.rng.WearLegacyScreenTest
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -42,9 +44,14 @@ class SectionedListHeaderFooterTest(
     @Test
     fun test() {
         runTest {
-            val columnState = ScalingLazyColumnDefaults.responsive().create()
+            val columnState = rememberResponsiveColumnState(
+                contentPadding = padding(
+                    first = ItemType.Text,
+                    last = ItemType.Chip,
+                ),
+            )
 
-            SectionedListPreview(columnState) {
+            ScreenScaffold(scrollState = columnState) {
                 SectionedList(columnState = columnState) {
                     section(state = sectionStateParam) {
                         header(visibleStates = headerVisibleStatesParam) { DownloadsHeader() }
